Add details to UI

+++ src/org/starlo/climate/CloudReport.java
@@ -2,5 +2,5 @@ package org.starlo.climate;
 
 public class CloudReport
 {
-    public Short all;
+    public Integer all;
 }
